Fixed plugin tests that relied on beforeResolve()
The problem with the tests is that they rely on beforeResolve() to add
a default dependency if none was specified before. However, when they
do this, the configuration is already in OBSERVED state, and change is
ignored after a deprecation warning.

Solution: introduced a new beforeObserve() callback, and used that in
the plugins. This will still break user implementations where
beforeResolved() is used.